Moved tests for updating bridge parameters into a separate file
