+++ solidity/test/bridge/Bridge.Wallets.test.ts
@@ -19,7 +19,6 @@ const { lastBlockTime, increaseTime } = helpers.time
 const fixture = async () => bridgeFixture()
 
 describe("Bridge - Wallets", () => {
-  let governance: SignerWithAddress
   let thirdParty: SignerWithAddress
 
   let walletRegistry: FakeContract<IWalletRegistry>
@@ -27,139 +26,9 @@ describe("Bridge - Wallets", () => {
 
   before(async () => {
     // eslint-disable-next-line @typescript-eslint/no-extra-semi
-    ;({ governance, thirdParty, walletRegistry, bridge } =
-      await waffle.loadFixture(fixture))
-  })
-
-  describe("updateWalletParameters", () => {
-    context("when caller is the contract owner", () => {
-      context("when all new parameter values are correct", () => {
-        const newWalletCreationPeriod = constants.walletCreationPeriod * 2
-        const newWalletMinBtcBalance = constants.walletMinBtcBalance.add(1000)
-        const newWalletMaxBtcBalance = constants.walletMaxBtcBalance.add(2000)
-        const newWalletMaxAge = constants.walletMaxAge * 2
-        const newWalletMaxBtcTransfer = constants.walletMaxBtcTransfer.add(1000)
-
-        let tx: ContractTransaction
-
-        before(async () => {
-          await createSnapshot()
-
-          tx = await bridge
-            .connect(governance)
-            .updateWalletParameters(
-              newWalletCreationPeriod,
-              newWalletMinBtcBalance,
-              newWalletMaxBtcBalance,
-              newWalletMaxAge,
-              newWalletMaxBtcTransfer
-            )
-        })
-
-        after(async () => {
-          await restoreSnapshot()
-        })
-
-        it("should set correct values", async () => {
-          const params = await bridge.walletParameters()
-
-          expect(params.walletCreationPeriod).to.be.equal(
-            newWalletCreationPeriod
-          )
-          expect(params.walletMinBtcBalance).to.be.equal(newWalletMinBtcBalance)
-          expect(params.walletMaxBtcBalance).to.be.equal(newWalletMaxBtcBalance)
-          expect(params.walletMaxAge).to.be.equal(newWalletMaxAge)
-          expect(params.walletMaxBtcTransfer).to.be.equal(
-            newWalletMaxBtcTransfer
-          )
-        })
-
-        it("should emit WalletParametersUpdated event", async () => {
-          await expect(tx)
-            .to.emit(bridge, "WalletParametersUpdated")
-            .withArgs(
-              newWalletCreationPeriod,
-              newWalletMinBtcBalance,
-              newWalletMaxBtcBalance,
-              newWalletMaxAge,
-              newWalletMaxBtcTransfer
-            )
-        })
-      })
-
-      context("when new minimum BTC balance is zero", () => {
-        it("should revert", async () => {
-          await expect(
-            bridge
-              .connect(governance)
-              .updateWalletParameters(
-                constants.walletCreationPeriod,
-                0,
-                constants.walletMaxBtcBalance,
-                constants.walletMaxAge,
-                constants.walletMaxBtcTransfer
-              )
-          ).to.be.revertedWith(
-            "Wallet minimum BTC balance must be greater than zero"
-          )
-        })
-      })
-
-      context(
-        "when new maximum BTC balance is not greater than the minimum",
-        () => {
-          it("should revert", async () => {
-            await expect(
-              bridge
-                .connect(governance)
-                .updateWalletParameters(
-                  constants.walletCreationPeriod,
-                  constants.walletMinBtcBalance,
-                  constants.walletMinBtcBalance,
-                  constants.walletMaxAge,
-                  constants.walletMaxBtcTransfer
-                )
-            ).to.be.revertedWith(
-              "Wallet maximum BTC balance must be greater than the minimum"
-            )
-          })
-        }
-      )
-
-      context("when new maximum BTC transfer is zero", () => {
-        it("should revert", async () => {
-          await expect(
-            bridge
-              .connect(governance)
-              .updateWalletParameters(
-                constants.walletCreationPeriod,
-                constants.walletMinBtcBalance,
-                constants.walletMaxBtcBalance,
-                constants.walletMaxAge,
-                0
-              )
-          ).to.be.revertedWith(
-            "Wallet maximum BTC transfer must be greater than zero"
-          )
-        })
-      })
-    })
-
-    context("when caller is not the contract owner", () => {
-      it("should revert", async () => {
-        await expect(
-          bridge
-            .connect(thirdParty)
-            .updateWalletParameters(
-              constants.walletCreationPeriod,
-              constants.walletMinBtcBalance,
-              constants.walletMaxBtcBalance,
-              constants.walletMaxAge,
-              constants.walletMaxBtcTransfer
-            )
-        ).to.be.revertedWith("Ownable: caller is not the owner")
-      })
-    })
+    ;({ thirdParty, walletRegistry, bridge } = await waffle.loadFixture(
+      fixture
+    ))
   })
 
   describe("requestNewWallet", () => {
